# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lecker Candy Store API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderRequestItemsItem } from "./createOrderRequestItemsItem";
import type { CreateOrderRequestPaymentMethod } from "./createOrderRequestPaymentMethod";

export interface CreateOrderRequest {
  customerPhone: string;
  customerName?: string;
  items: CreateOrderRequestItemsItem[];
  notes?: string;
  deliveryAddress?: string;
  paymentMethod: CreateOrderRequestPaymentMethod;
}