# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lecker Candy Store API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductRequest {
  name: string;
  nameAr: string;
  category: string;
  price: number;
  description?: string;
  imageUrl?: string;
  isActive?: boolean;
  sortOrder?: number;
}