# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lecker Candy Store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  userId?: number | null;
  customerPhone: string;
  customerName?: string | null;
  status: OrderStatus;
  total: number;
  items: OrderItem[];
  notes?: string | null;
  deliveryAddress?: string | null;
  paymentMethod: OrderPaymentMethod;
  createdAt: Date;
  updatedAt: Date;
}